+++ addons/mass_mailing/__manifest__.py
@@ -36,16 +36,17 @@
         'wizard/mailing_list_merge_views.xml',
         'wizard/mailing_mailing_test_views.xml',
         'wizard/mailing_mailing_schedule_date_views.xml',
+        'report/mailing_trace_report_views.xml',
         'views/mailing_filter_views.xml',
-        'views/mailing_mailing_views_menus.xml',
         'views/mailing_trace_views.xml',
         'views/link_tracker_views.xml',
         'views/mailing_contact_views.xml',
+        'views/mailing_contact_subscription_views.xml',
         'views/mailing_list_views.xml',
         'views/mailing_mailing_views.xml',
         'views/res_config_settings_views.xml',
         'views/utm_campaign_views.xml',
-        'report/mailing_trace_report_views.xml',
+        'views/mailing_menus.xml',
         'views/assets.xml',
         'views/mass_mailing_templates_portal.xml',
         'views/themes_templates.xml',
