+++ addons/mass_mailing/models/__init__.py
@@ -3,6 +3,7 @@
 
 from . import ir_model
 from . import link_tracker
+from . import mailing_contact_subscription
 from . import mailing_contact
 from . import mailing_list
 from . import mailing_trace
