+++ addons/mass_mailing/report/mailing_trace_report_views.xml
@@ -95,8 +95,4 @@
 </p>
 </field>
        </record>
-
-       <menuitem name="Reporting" id="menu_mass_mailing_report" sequence="99"
-            parent="mass_mailing_menu_root"
-            action="mailing_trace_report_action_mail"/>
 </odoo>
