+++ addons/mass_mailing/views/link_tracker_views.xml
@@ -78,12 +78,4 @@
             </xpath>
         </field>
     </record>
-
-    <!-- MENU TO HANLDE LINK DATA IN MM -->
-    <menuitem id="link_tracker_menu_mass_mailing"
-        name="Link Tracker"
-        parent="mass_mailing_configuration"
-        sequence="5"
-        action="link_tracker.link_tracker_action"/>
-
 </odoo>
