+++ addons/mass_mailing/views/mailing_list_views.xml
@@ -293,8 +293,4 @@
           </p>
         </field>
     </record>
-
-    <menuitem name="Mailing Lists" id="menu_email_mass_mailing_lists"
-        parent="mass_mailing_mailing_list_menu" sequence="3"
-        action="action_view_mass_mailing_lists"/>
 </odoo>
