+++ addons/mass_mailing/views/mailing_mailing_views.xml
@@ -718,10 +718,4 @@
             <field name="res_model">mailing.mailing</field>
             <field name="view_mode">form</field>
         </record>
-
-        <menuitem name="Mailings" id="mass_mailing_menu"
-            parent="mass_mailing_menu_root"
-            sequence="1"
-            action="mailing_mailing_action_mail"/>
-
 </odoo>
