--- addons/mass_mailing/views/mailing_mailing_views_menus.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <!-- Marketing / Mailing -->
-    <menuitem name="Email Marketing" id="mass_mailing_menu_root" sequence="115" web_icon="mass_mailing,static/description/icon.svg" groups="mass_mailing.group_mass_mailing_user"/>
-    <menuitem name="Mailing Lists" id="mass_mailing_mailing_list_menu"
-        parent="mass_mailing_menu_root" sequence="2"/>
-
-    <!-- Marketing / Configuration -->
-    <menuitem name="Configuration" id="mass_mailing_configuration"
-        parent="mass_mailing_menu_root"
-        sequence="100"/>
-
-    <!-- Configuration / Blacklist -->
-    <menuitem id="mail_blacklist_mm_menu" name="Blacklisted Email Addresses"
-        action="mail.mail_blacklist_action"
-        parent="mass_mailing_configuration"/>
-
-    <!-- Configuration / Favorite Filters -->
-    <menuitem id="mailing_filter_menu_action"
-        action="mailing_filter_action"
-        parent="mass_mailing_configuration"/>
-
-    <!-- Technical / Mass Mailing -->
-    <menuitem id="mailing_mailing_menu_technical"
-        name="Mass Mailing"
-        sequence="4"
-        parent="base.menu_custom"/>
-</odoo>
