+++ addons/mass_mailing/views/mailing_trace_views.xml
@@ -140,10 +140,4 @@
             </p>
         </field>
     </record>
-
-    <!-- Add in Technical/Email -->
-    <menuitem id="menu_email_statistics"
-        name="Mailing Traces"
-        parent="mass_mailing.mailing_mailing_menu_technical" sequence="2"
-        action="mailing_trace_action"/>
 </odoo>
