+++ addons/mass_mailing/views/res_config_settings_views.xml
@@ -78,7 +78,4 @@
             <field name="target">inline</field>
             <field name="context">{'module' : 'mass_mailing', 'bin_size': False}</field>
         </record>
-
-        <menuitem id="menu_mass_mailing_global_settings" name="Settings"
-            parent="mass_mailing_configuration" sequence="0" action="action_mass_mailing_configuration" groups="base.group_system"/>
 </odoo>
