+++ addons/mass_mailing_sms/__manifest__.py
@@ -15,15 +15,14 @@
     'data': [
         'data/utm_data.xml',
         'security/ir.model.access.csv',
-        'views/mailing_sms_menus.xml',
+        'report/mailing_trace_report_views.xml',
         'views/mailing_list_views.xml',
         'views/mailing_contact_views.xml',
         'views/mailing_trace_views.xml',
         'views/mailing_mailing_views.xml',
-        'views/link_tracker_views.xml',
         'views/mass_mailing_sms_templates_portal.xml',
         'views/utm_campaign_views.xml',
-        'report/mailing_trace_report_views.xml',
+        'views/mailing_sms_menus.xml',
         'wizard/sms_composer_views.xml',
         'wizard/mailing_sms_test_views.xml',
     ],
