+++ addons/mass_mailing_sms/report/mailing_trace_report_views.xml
@@ -83,9 +83,4 @@
         <field name="act_window_id" ref="mailing_trace_report_action_sms"/>
         <field name="view_id" ref="mailing_trace_report_sms_view_tree"/>
     </record>
-
-    <!-- SMS Marketing / Reporting -->
-    <record id="mass_mailing_sms_menu_reporting" model="ir.ui.menu">
-        <field name="action" ref="mailing_trace_report_action_sms"/>
-    </record>
 </odoo>
