--- addons/mass_mailing_sms/views/link_tracker_views.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <menuitem id="link_tracker_menu"
-        name="Link Tracker"
-        parent="mass_mailing_sms_menu_configuration"
-        sequence="2"
-        action="link_tracker.link_tracker_action"/>
-</odoo>
