+++ addons/mass_mailing_sms/views/mailing_contact_views.xml
@@ -91,10 +91,4 @@
             </p>
         </field>
     </record>
-
-    <!-- SMS Marketing / Contacts Lists / Contacts Lists -->
-    <record id="mass_mailing_sms.mailing_contact_menu_sms" model="ir.ui.menu">
-        <field name="action" ref="mailing_contact_action_sms"/>
-    </record>
-
 </odoo>
