+++ addons/mass_mailing_sms/views/mailing_list_views.xml
@@ -42,9 +42,4 @@
           </p>
         </field>
     </record>
-
-    <!-- SMS Marketing / Contacts Lists / Contacts Lists -->
-    <record id="mass_mailing_sms.mailing_list_menu_sms" model="ir.ui.menu">
-        <field name="action" ref="mailing_list_action_sms"/>
-    </record>
 </odoo>
