+++ addons/mass_mailing_sms/views/mailing_mailing_views.xml
@@ -281,9 +281,4 @@
         <field name="act_window_id" ref="mailing_mailing_action_sms"/>
     </record>
 
-    <!-- SMS Marketing / SMS Marketing -->
-    <record id="mass_mailing_sms_menu_mass_sms" model="ir.ui.menu">
-        <field name="action" ref="mailing_mailing_action_sms"/>
-    </record>
-
 </data></odoo>
