+++ addons/mass_mailing_sms/views/utm_campaign_views.xml
@@ -58,9 +58,4 @@
             </xpath>
         </field>
     </record>
-
-    <menuitem name="Campaigns" id="menu_email_campaigns"
-        parent="mass_mailing_sms_menu_root" sequence="5"
-        action="mass_mailing.action_view_utm_campaigns"
-        groups="mass_mailing.group_mass_mailing_campaign"/>
 </odoo>
